Major cleanups.
Make the code build on more platforms and be warning clean. This also
includes the python bindings.

fileIndexer::processNgrams now returns void. The bool pointer was never used.

Move the search prototype out of the extern C block. This caused clang to
be unhappy on OS X 10.9.

PyList_Size() returns a Py_ssize_t, not an int.

Cast the return value in cpu_count(). This code is currently unused but
as we have plans to make this parallel we will use it eventually.

Don't include relative paths as they are given in the Makefile.

Silence a bunch of warnings due to missing headers for read(2), write(2)
and related functions.

Be specific with casting to mmap(3), munmap(3) and just all over the place in
general. This, and the lack of (char *) casting was probably the biggest
source of warnings.

Remove unused variables and the unused version of fileIndexer::processFile().

Whitespaces changes while I'm here.

+++ snugglefish.cpp
@@ -81,14 +81,14 @@ uint32_t cpu_count(){
              << strerror(errno) << endl;
         return 1;
     }
-    return procs;
+    return (uint32_t) procs;
 }
 
 
 
 void handler(int sig) {
   void *array[10];
-  size_t size;
+  int size;
 
   // get void*'s for all entries on the stack
   size = backtrace(array, 10);
